Fix Issue related to unique identity generator and added support to have short url via get method as well

+++ src/main/java/com/prakash/docker/urlshortening/service/URLShortService.java
@@ -20,7 +20,7 @@ public class URLShortService {
 	
 	
 	public String createShortUrl(UrlShortRequest request) {
-		int number = uniqueSequenceGenerator.getUniqueBase10Number();
+		Long number = uniqueSequenceGenerator.getUniqueBase10Number();
 		String shortUrl=urlGenerator.encode(number);
 		request.setId(number);
 		request.setShortUrl(shortUrl);
